fixed typo

+++ packages/contracts/contracts/TestContracts/CDPManagerTester.sol
@@ -9,23 +9,23 @@ for testing the parent's internal functions. */
 
 contract TroveManagerTester is TroveManager {
 
-    function computeICR(uint _coll, uint _debt, uint _price) external pure returns (uint) {
+    function computeICR(uint256 _coll, uint256 _debt, uint256 _price) external pure returns (uint256) {
         return LiquityMath._computeCR(_coll, _debt, _price);
     }
 
-    function getCollGasCompensation(uint _coll) external pure returns (uint) {
+    function getCollGasCompensation(uint256 _coll) external pure returns (uint256) {
         return _getCollGasCompensation(_coll);
     }
 
-    function getLUSDGasCompensation() external pure returns (uint) {
+    function getLUSDGasCompensation() external pure returns (uint256) {
         return LUSD_GAS_COMPENSATION;
     }
 
-    function getCompositeDebt(uint _debt) external pure returns (uint) {
+    function getCompositeDebt(uint256 _debt) external pure returns (uint256) {
         return _getCompositeDebt(_debt);
     }
 
-    function unprotectedDecayBaseRateFromBorrowing() external returns (uint) {
+    function unprotectedDecayBaseRateFromBorrowing() external returns (uint256) {
         baseRate = _calcDecayedBaseRate();
         assert(baseRate >= 0 && baseRate <= DECIMAL_PRECISION);
         
@@ -33,7 +33,7 @@ contract TroveManagerTester is TroveManager {
         return baseRate;
     }
 
-    function minutesPassedSinceLastFeeOp() external view returns (uint) {
+    function minutesPassedSinceLastFeeOp() external view returns (uint256) {
         return _minutesPassedSinceLastFeeOp();
     }
 
@@ -41,20 +41,20 @@ contract TroveManagerTester is TroveManager {
         lastFeeOperationTime = block.timestamp;
     }
 
-    function setBaseRate(uint _baseRate) external {
+    function setBaseRate(uint256 _baseRate) external {
         baseRate = _baseRate;
     }
 
-    function callGetRedemptionFee(uint _ETHDrawn) external view returns (uint) {
+    function callGetRedemptionFee(uint256 _ETHDrawn) external view returns (uint256) {
         return _getRedemptionFee(_ETHDrawn);
     }  
 
-    function getActualDebtFromComposite(uint _debtVal) external pure returns (uint) {
+    function getActualDebtFromComposite(uint256 _debtVal) external pure returns (uint256) {
         return _getNetDebt(_debtVal);
     }
 
     function callInternalRemoveTroveOwner(address _troveOwner) external {
-        uint troveOwnersArrayLength = TroveOwners.length;
+        uint256 troveOwnersArrayLength = TroveOwners.length;
         _removeTroveOwner(_troveOwner, troveOwnersArrayLength);
     }
 }
